+++ lib/components/datatypes/servant.dart
@@ -350,7 +350,7 @@ class ServantBaseInfo {
       localizeNoun(illustrator, illustratorJp, illustratorEn);
 
   List<String> get lCV =>
-      localizeNoun<List<String>>(cv, cvJp, cvEn, () => <String>[]);
+      localizeNoun<List<String>>(cv, cvJp, cvEn, k: () => <String>[]);
 
   factory ServantBaseInfo.fromJson(Map<String, dynamic> data) =>
       _$ServantBaseInfoFromJson(data);
@@ -461,7 +461,7 @@ class SvtProfileData {
   });
 
   String get lDescription =>
-      localizeNoun(description, descriptionJp, descriptionEn, () => '???');
+      localizeNoun(description, descriptionJp, descriptionEn, k: () => '???');
 
   String get lCondition => localizeNoun(condition, null, conditionEn);