+++ lib/modules/craft/craft_detail_page.dart
@@ -1,4 +1,6 @@
+import 'package:auto_size_text/auto_size_text.dart';
 import 'package:chaldea/components/components.dart';
+import 'package:chaldea/modules/shared/lang_switch.dart';
 import 'package:chaldea/modules/summon/summon_detail_page.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -14,7 +16,7 @@ class CraftDetailPage extends StatefulWidget {
 }
 
 class _CraftDetailPageState extends State<CraftDetailPage> {
-  bool useLangCn = false;
+  Language? lang;
 
   late CraftEssence ce;
 
@@ -22,7 +24,6 @@ class _CraftDetailPageState extends State<CraftDetailPage> {
   void initState() {
     super.initState();
     ce = widget.ce;
-    useLangCn = Language.isCN;
   }
 
   @override
@@ -32,7 +33,8 @@ class _CraftDetailPageState extends State<CraftDetailPage> {
     return Scaffold(
       appBar: AppBar(
         leading: BackButton(),
-        title: Text(ce.localizedName),
+        title: AutoSizeText(ce.localizedName, maxLines: 1),
+        titleSpacing: 0,
         actions: [
           IconButton(
             tooltip:
@@ -46,8 +48,8 @@ class _CraftDetailPageState extends State<CraftDetailPage> {
             icon: status == 1
                 ? Icon(Icons.favorite)
                 : status == 2
-                    ? Icon(Icons.favorite, color: Colors.redAccent)
-                    : Icon(Icons.favorite_outline),
+                ? Icon(Icons.favorite, color: Colors.redAccent)
+                : Icon(Icons.favorite_outline),
           ),
           _popupButton,
         ],
@@ -56,27 +58,18 @@ class _CraftDetailPageState extends State<CraftDetailPage> {
         children: <Widget>[
           Expanded(
               child: CraftDetailBasePage(
-            ce: ce,
-            useLangCn: useLangCn,
+                ce: ce,
+            lang: lang,
             showSummon: true,
           )),
           ButtonBar(alignment: MainAxisAlignment.center, children: [
-            ToggleButtons(
-              constraints: BoxConstraints(),
-              selectedColor: Colors.white,
-              fillColor: Theme.of(context).primaryColor,
-              onPressed: (i) {
+            ProfileLangSwitch(
+              primary: lang,
+              onChanged: (v) {
                 setState(() {
-                  useLangCn = i == 0;
+                  lang = v;
                 });
               },
-              children: List.generate(
-                  2,
-                  (i) => Padding(
-                        padding: EdgeInsets.all(6),
-                        child: Text(['中', '日'][i]),
-                      )),
-              isSelected: List.generate(2, (i) => useLangCn == (i == 0)),
             ),
             for (var i = 0; i < 2; i++)
               ElevatedButton(
@@ -134,14 +127,11 @@ class _CraftDetailPageState extends State<CraftDetailPage> {
 
 class CraftDetailBasePage extends StatelessWidget {
   final CraftEssence ce;
-  final bool useLangCn;
+  final Language? lang;
   final bool showSummon;
 
   const CraftDetailBasePage(
-      {Key? key,
-      required this.ce,
-      this.useLangCn = false,
-      this.showSummon = false})
+      {Key? key, required this.ce, this.lang, this.showSummon = false})
       : super(key: key);
 
   @override
@@ -153,7 +143,7 @@ class CraftDetailBasePage extends StatelessWidget {
           CustomTableRow(children: [
             TableCellData(
               child:
-                  Text(ce.name, style: TextStyle(fontWeight: FontWeight.bold)),
+              Text(ce.name, style: TextStyle(fontWeight: FontWeight.bold)),
               isHeader: true,
             )
           ]),
@@ -228,10 +218,7 @@ class CraftDetailBasePage extends StatelessWidget {
           ]),
           CustomTableRow(children: [
             TableCellData(
-              child: Text(
-                  Localized.craftFilter.of(ce.category) +
-                      ' - ' +
-                      ce.categoryText,
+              child: Text(Localized.craftFilter.of(ce.category),
                   textAlign: TextAlign.center),
             )
           ]),
@@ -294,7 +281,11 @@ class CraftDetailBasePage extends StatelessWidget {
           CustomTableRow(
             children: [
               TableCellData(
-                text: ce.lDescription,
+                text: LocalizedText(
+                        chs: ce.description ?? '???',
+                        jpn: ce.descriptionJp,
+                        eng: ce.descriptionEn)
+                    .ofPrimary(lang ?? Language.current),
                 alignment: Alignment.centerLeft,
                 padding: EdgeInsets.symmetric(horizontal: 10, vertical: 8),
               )
@@ -307,23 +298,23 @@ class CraftDetailBasePage extends StatelessWidget {
             CustomTableRow(children: [
               TableCellData(
                   child: Column(
-                mainAxisSize: MainAxisSize.min,
-                children: [
-                  for (var summon in summons)
-                    ListTile(
-                      title: Text(summon.localizedName, maxLines: 1),
-                      dense: true,
-                      visualDensity: VisualDensity.compact,
-                      onTap: () {
-                        SplitRoute.push(
-                          context: context,
-                          builder: (context, _) =>
-                              SummonDetailPage(summon: summon),
-                        );
-                      },
-                    )
-                ],
-              ))
+                    mainAxisSize: MainAxisSize.min,
+                    children: [
+                      for (var summon in summons)
+                        ListTile(
+                          title: Text(summon.localizedName, maxLines: 1),
+                          dense: true,
+                          visualDensity: VisualDensity.compact,
+                          onTap: () {
+                            SplitRoute.push(
+                              context: context,
+                              builder: (context, _) =>
+                                  SummonDetailPage(summon: summon),
+                            );
+                          },
+                        )
+                    ],
+                  ))
             ])
           ]
         ],
@@ -335,7 +326,7 @@ class CraftDetailBasePage extends StatelessWidget {
     if (characters.isEmpty) return '-';
     return characters.map((e) {
       final svt =
-          db.gameData.servants.values.firstWhereOrNull((s) => s.mcLink == e);
+      db.gameData.servants.values.firstWhereOrNull((s) => s.mcLink == e);
       return svt?.info.localizedName ?? e;
     }).join(', ');
   }